feat: set up wifi on start

+++ firmware/lib/indicator.h
@@ -33,7 +33,7 @@ public:
     }
   }
 
-  void setLevel(byte value)
+  void setLevel(int value)
   {
     int pwm = constrain(map(value, 0, 100, 0, 255), 0, 255);
     analogWrite(pinIndicator, pwm);
